Put yarn install back into docker compose file.
The yarn install command must run from the compose yaml file because during container creation the mount point is not set.
We mount the docker image with the source directory of the app in order to keep updates in sync with the docker image without rebuilding the image after each change of the source code.